allow arbitrary table batch size overrides for reporting

+++ api/src/main/java/org/pmiops/workbench/reporting/ReportingTableParams.java
@@ -8,7 +8,7 @@ import org.pmiops.workbench.reporting.insertion.InsertAllRequestPayloadTransform
 
 public record ReportingTableParams<T extends ReportingBase>(
     String bqTableName,
-    long batchSize,
+    int batchSize,
     InsertAllRequestPayloadTransformer<T> bqInsertionBuilder,
     BiFunction<Long, Long, List<T>> rwbBatchQueryFn,
     IntSupplier rwbTableCountFn) {}
